Обновлен файл docker-compose.yml для интеграции Redis и Celery в AI сервисы. Добавлены конфигурации для Redis и Flower, а также создан новый файл redis.conf с настройками. Реализованы тесты для асинхронного выполнения задач в test_docker_celery.py. Обновлены файлы ai_services для поддержки новой архитектуры с использованием Celery. Исправлены модели данных для поддержки BigInteger диапазона в post.py. Обновлен ROADMAP_v2.md с новыми целями и задачами для миграции на Docker.

+++ backend/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey, JSON, Boolean, Float, UniqueConstraint, Index
+from sqlalchemy import Column, Integer, BigInteger, String, Text, DateTime, ForeignKey, JSON, Boolean, Float, UniqueConstraint, Index
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
@@ -60,8 +60,8 @@ class PostCache(Base):
     __tablename__ = "posts_cache"
     
     id = Column(Integer, primary_key=True, index=True)
-    channel_telegram_id = Column(Integer, nullable=False, index=True)
-    telegram_message_id = Column(Integer, nullable=False)
+    channel_telegram_id = Column(BigInteger, nullable=False, index=True)
+    telegram_message_id = Column(BigInteger, nullable=False)
     title = Column(Text)
     content = Column(Text)
     # Условные типы данных в зависимости от БД
